Refactor: sessionToken (param) & apiPicker (module)
Refactor:
- made sessionToken as parameter for functional components

- made aapiPicker as module
