+++ src/hooks/useStudents.ts
@@ -4,6 +4,7 @@
 import { useState, useEffect } from "react";
 import type { StudentRecord } from "../utils/dataTypes";
 
+import { apiPicker } from "../services/apiPicker";
 interface UseStudentsResult {
   students: StudentRecord[];
   loading: boolean;
@@ -11,44 +12,34 @@ interface UseStudentsResult {
   refetchStudents: () => void; // Add a refetch function
 }
 
-// Ensure VITE_API_PY_URL is set in your .env file (e.g., VITE_API_PY_URL=http://localhost:5000/api/students)
-// const apiURL = import.meta.env.VITE_API_PY_URL_OTHERHOST;
-const apiPicker = (theChoice: number, theToken: string) => {
-  const choice1 = import.meta.env.VITE_API_PY_URL;
-  const choice2 = import.meta.env.VITE_API_GO_URL;
-
-  if (theChoice === 1) {
-    return choice1;
-  } else {
-    return String(choice2 + "?teacherID=" + theToken);
-  }
-};
-// props.theUserID
-
+// export const useStudents = (
+//   theChoice: number,
+//   theUserID: string
+// ): UseStudentsResult => {
 export const useStudents = (
   theChoice: number,
-  theUserID: string
+  theSessionToken: string
 ): UseStudentsResult => {
   const [students, setStudents] = useState<StudentRecord[]>([]);
   const [loading, setLoading] = useState<boolean>(true);
   const [error, setError] = useState<string | null>(null);
   const [triggerRefetch, setTriggerRefetch] = useState(0); // State to trigger refetch
 
-  const apiURL = apiPicker(theChoice, theUserID);
+  // const apiURL = apiPicker(theChoice, theUserID);
+  const apiURL = apiPicker(theChoice);
 
-  const headers = {};
-
+  // CHQ: Gemini AI corrected the headers assignment
+  const headers = {
+    Authorization: `Bearer ${theSessionToken}`,
+  };
   // const apiURL=apiPicker(2)
 
   const fetchStudents = async () => {
     setLoading(true); // Set loading to true on every fetch attempt
     setError(null); // Clear any previous errors
 
-    console.log("theUserID is:");
-    console.log(theUserID);
-
-    console.log("userId is:");
-    console.log(theUserID);
+    // console.log("theUserID is:");
+    // console.log(theUserID);
 
     if (!apiURL) {
       setError("api url is not defined in environment variables.");
