+++ src/utils/dataTypes.ts
@@ -30,36 +30,46 @@ export interface StudentRecord {
   enrollment_date?: string;
 }
 
+// CHQ: Gemini AI generated interface UserData
+export interface UserData {
+  loginsIds: string[];
+  userId: string;
+  userNames: {
+    name: string;
+    email: string;
+    phone: string;
+    verifiedEmail: boolean;
+    verifiedPhone: boolean;
+    roleNames: string[];
+    logins: any[];
+  }[];
+  userTenants: any[];
+  status: string;
+  OAuth: {
+    google: boolean;
+  };
+  SAML: boolean;
+  SCIM: boolean;
+  TOTP: boolean;
+  createTime: number;
+  customAttributes: {};
+  email: string;
+  externalIds: string[];
+  familyName: string;
+  givenName: string;
+  loginIds: string[];
+  middleName: string;
+  password: boolean;
+  phone: string;
+  picture: string;
+  roleNames: string[];
+  ssoIds: any[];
+  test: boolean;
+  verifiedEmail: boolean;
+  verifiedPhone: boolean;
+  webauthn: boolean;
+}
 export interface Item {
   id: number;
   value: string;
-}
-
-export interface FoodItem {
-  id: number;
-  value: string;
-}
-
-export interface QtyItem {
-  id: number;
-  value: number;
-}
-
-// export interface QtyItem {
-//   id?: number;
-//   value: number;
-// }
-
-export interface ProcessedLine {
-  id: string; // Unique ID for React key
-  text: string;
-  qty: number;
-  matchedItem: FoodItem | null; // The matched food item
-  originalMatchConfidence?: number; // Optional: confidence score for matching
-}
-
-export interface TextProcessorProps {
-  foodItems: FoodItem[]; // The list of all available food items from MyTable
-  onProcessedLinesChange: (lines: ProcessedLine[]) => void; // Callback to pass processed lines to parent
-  // onDataStructure: (theData: Map<string, string[]>) => void;
 }
